+++ packages/grafana-schema/package.json
@@ -36,10 +36,12 @@
     "postpack": "mv package.json.bak package.json"
   },
   "devDependencies": {
+    "@grafana/data": "workspace:^",
     "@grafana/tsconfig": "^1.2.0-rc1",
     "@rollup/plugin-commonjs": "23.0.2",
     "@rollup/plugin-json": "5.0.1",
     "@rollup/plugin-node-resolve": "15.1.0",
+    "@types/react": "^18.2.14",
     "esbuild": "0.17.19",
     "glob": "^10.2.7",
     "rimraf": "4.4.0",
